trying to stop MemoryErrors in param_WUE.py

+++ src/param_WUE.py
@@ -53,7 +53,7 @@ def zipUpNetCDFFile(netcdfFile):
 engine = win32com.client.Dispatch('DAO.DBEngine.120')
 db = engine.OpenDatabase('D:/Users/andrewcottam/Documents/fao_ftp_files.accdb')
 queryDef = db.CreateQueryDef("", "SELECT distinct fullpath, archivename,filename FROM DownloadedZipFiles WHERE (((DownloadedZipFiles.filename) Like '*_yield_sea_*'));")
-# queryDef = db.CreateQueryDef("", "SELECT distinct fullpath, archivename,filename FROM DownloadedZipFiles WHERE filename='ClimAf_1_1960-2100_lpjg_can_wfdei_qmbc_B1_yield_sea_trpe_rf.nc.gz';")
+# queryDef = db.CreateQueryDef("", "SELECT distinct fullpath, archivename,filename FROM DownloadedZipFiles WHERE filename'ClimAf_1_1960-2100_lpjg_can_wfdei_qmbc_B2_yield_sea_trmi_rf.nc.gz';")
 r = queryDef.OpenRecordset()
 total = r.RecordCount
 counter = 1
@@ -63,34 +63,37 @@ while not r.EOF:
     gzfile1 = str(r.fullpath)
     gzfile2 = gzfile1.replace("_yield_sea_", "_etotx_sea_")
     archivename1 = str(r.archivename)
-    rightpart = archivename1[archivename1.find("_yield_") - 3:].replace("_yield_sea_", "_etotx_sea_")  # archive name may not be the same as yield one - e.g. somd
-    leftbit = archivename1[:archivename1.find("_wfdei_")]
-    sql = "Select archivename from downloadedzipfiles where filename like '*_etotx_sea_*' and filename like '*" + rightpart + "*' and filename like '*" + leftbit + "*';"
-    querydef2 = db.CreateQueryDef("", sql)
-    q2 = querydef2.OpenRecordset()
-    if q2.EOF == False:
-        archivename2 = str(q2.archivename)
-        try:
-            print "Unzipping " + gzfile1
-            UnzipGZFile(gzfile1, archivename1)
-            print "Unzipping " + gzfile2
-            UnzipGZFile(gzfile2, archivename2)
-            print "Creating the Netcdf file"
-            netcdfFile = CreateNetCDF(archivename1, archivename2)
-            print "Created " + netcdfFile + "\nZipping file"
-            zippedNetCDF = zipUpNetCDFFile(netcdfFile)
-            print "Finished\n\n"
-            os.remove(netcdfFile)
-        except (IOError):
+    if (os.path.exists(INPUT_ZIPS+archivename1.replace("yield","wuexx")+".gz")):
+        print "Output zipped NetCDF file already exists"
+    else:
+        rightpart = archivename1[archivename1.find("_yield_") - 3:].replace("_yield_sea_", "_etotx_sea_")  # archive name may not be the same as yield one - e.g. somd
+        leftbit = archivename1[:archivename1.find("_wfdei_")]
+        sql = "Select archivename from downloadedzipfiles where filename like '*_etotx_sea_*' and filename like '*" + rightpart + "*' and filename like '*" + leftbit + "*';"
+        querydef2 = db.CreateQueryDef("", sql)
+        q2 = querydef2.OpenRecordset()
+        if q2.EOF == False:
+            archivename2 = str(q2.archivename)
+            try:
+                print "Unzipping " + gzfile1
+                UnzipGZFile(gzfile1, archivename1)
+                print "Unzipping " + gzfile2
+                UnzipGZFile(gzfile2, archivename2)
+                print "Creating the Netcdf file"
+                netcdfFile = CreateNetCDF(archivename1, archivename2)
+                print "Created " + netcdfFile + "\nZipping file"
+                zippedNetCDF = zipUpNetCDFFile(netcdfFile)
+                print "Finished\n\n"
+                os.remove(netcdfFile)
+            except (IOError):
+                logging.error("Error creating NetCDF file for WUE parameter. Required gz file " + gzfile2 + " not found")
+                print "Error creating NetCDF file for WUE parameter. Required gz file " + gzfile2 + " not found"
+            if os.path.exists(INPUT_DATA + archivename2):
+                os.remove(INPUT_DATA + archivename2)
+        else:
             logging.error("Error creating NetCDF file for WUE parameter. Required gz file " + gzfile2 + " not found")
             print "Error creating NetCDF file for WUE parameter. Required gz file " + gzfile2 + " not found"
-        if os.path.exists(INPUT_DATA + archivename2):
-            os.remove(INPUT_DATA + archivename2)
-    else:
-        logging.error("Error creating NetCDF file for WUE parameter. Required gz file " + gzfile2 + " not found")
-        print "Error creating NetCDF file for WUE parameter. Required gz file " + gzfile2 + " not found"
-    if os.path.exists(INPUT_DATA + archivename1):
-        os.remove(INPUT_DATA + archivename1)
+        if os.path.exists(INPUT_DATA + archivename1):
+            os.remove(INPUT_DATA + archivename1)
     r.MoveNext()
     counter = counter + 1
